[#36] refactor: 영수증 공유 페이지 컴포넌트 분리
